+++ src/server/game-data/parties.ts
@@ -2,7 +2,7 @@ import { Version } from "../routes/versions";
 import { TemporaryUpdateTimeline } from "./changes";
 import { FileRef } from "./files";
 import { RoomName } from "./rooms";
-import { ANNIVERSARY_5_START, CAVE_EXPEDITION_END, CAVE_OPENING_END, CAVE_OPENING_START, CHRISTMAS_2005_ENDS, CHRISTMAS_2007_START, DIG_OUT_DOJO_END, EARTH_DAY_2010_END, EARTH_DAY_2010_START, EGG_HUNT_2006_END, EGG_HUNT_2006_START, EPF_RELEASE, FAIR_2010_START, FIRE_CELEBRATION_START, HALLOWEEN_2010_START, LIGHTHOUSE_PARTY_START, MUSIC_JAM_08_START, MUSIC_JAM_2010_CONST_START, MUSIC_JAM_2010_START, NEW_YEARS_2010_UPDATE, PET_SHOP_RELEASE, PIZZA_PARLOR_OPENING_END, PIZZA_PARLOR_OPENING_START, PLANET_Y_2010, PUFFLE_PARTY_10_CONST_START, ROCKHOPPER_ARRIVAL_END, ROCKHOPPER_ARRIVAL_PARTY_START, SPORT_PARTY_END, SPORT_PARTY_START, STADIUM_GAMES_END, SUMMER_PARTY_END, SUMMER_PARTY_START, WATER_CELEBRATION_END, WATER_HUNT_END, WINTER_FIESTA_08_START } from "./updates";
+import { ANNIVERSARY_5_START, CAVE_EXPEDITION_END, CAVE_OPENING_END, CAVE_OPENING_START, CHRISTMAS_2005_ENDS, CHRISTMAS_2007_START, DIG_OUT_DOJO_END, EARTH_DAY_2010_END, EARTH_DAY_2010_START, EGG_HUNT_2006_END, EGG_HUNT_2006_START, EPF_RELEASE, FAIR_2010_START, FIRE_CELEBRATION_START, FIRE_CONST_START, HALLOWEEN_2010_START, LIGHTHOUSE_PARTY_START, MUSIC_JAM_08_START, MUSIC_JAM_2010_CONST_START, MUSIC_JAM_2010_START, NEW_YEARS_2010_UPDATE, PET_SHOP_RELEASE, PIZZA_PARLOR_OPENING_END, PIZZA_PARLOR_OPENING_START, PLANET_Y_2010, PUFFLE_PARTY_10_CONST_START, ROCKHOPPER_ARRIVAL_END, ROCKHOPPER_ARRIVAL_PARTY_START, SPORT_PARTY_END, SPORT_PARTY_START, STADIUM_GAMES_END, SUMMER_PARTY_END, SUMMER_PARTY_START, WATER_CELEBRATION_END, WATER_HUNT_END, WINTER_FIESTA_08_START } from "./updates";
 
 // room name -> file Id
 export type RoomChanges = Partial<Record<RoomName, FileRef>>;
@@ -26,6 +26,10 @@ export type PartyChanges = {
   generalChanges?: Record<string, FileRef>;
 
   roomComment?: string;
+
+  map?: FileRef;
+
+  music?: Partial<Record<RoomName, number>>;
 }
 
 export type Party = PartyChanges & {
@@ -38,7 +42,6 @@ export type Party = PartyChanges & {
   startComment?: string;
   endComment?: string;
 
-  music?: Partial<Record<RoomName, number>>;
   construction?: Construction;
   /** Scavenger Hunt icon is loaded by the dependency, must be specified */
   scavengerHunt2010?: {
@@ -190,8 +193,7 @@ export const PARTIES: TemporaryUpdateTimeline<Party, PartyChanges> = [
         // now has path to the plaza
         forts: 'archives:ArtworkRoomsForts12.swf'
       },
-      // TODO map abstraction in general changes?
-      generalChanges: { 'artwork/maps/island5.swf': 'approximation:map_plaza_no_berg.swf' }
+      map: 'approximation:map_plaza_no_berg.swf'
     }
   },
   {
@@ -2047,7 +2049,7 @@ export const PARTIES: TemporaryUpdateTimeline<Party, PartyChanges> = [
       updates: [
         {
           comment: 'Construction for the Fire Dojo begins',
-          date: '2009-11-13',
+          date: FIRE_CONST_START,
           changes: {
             dojofire: 'archives:RoomsDojofire-FireCelebratePre.swf',
             // reverting back to normality
@@ -2851,8 +2853,8 @@ export const PARTIES: TemporaryUpdateTimeline<Party, PartyChanges> = [
       'dojohide': 'archives:HalloweenParty2010DojoHide.swf',
       coffee: 'archives:HalloweenParty2010Coffee.swf'
     },
+    map: 'archives:HalloweenParty2010Map.swf',
     globalChanges: {
-      'content/map.swf': 'archives:HalloweenParty2010Map.swf',
       'scavenger_hunt/hunt_ui.swf': ['recreation:halloween_hunt_2010.swf', 'hunt_ui', 'halloween_hunt'],
       'binoculars/empty.swf': 'archives:ContentBinoculars-HalloweenParty2007.swf', // from 2007 party
       'igloo/assets/igloo_background.swf': 'archives:Igloo_background_nightstorm.swf', // from 2011 party
@@ -3072,9 +3074,7 @@ export const PARTIES: TemporaryUpdateTimeline<Party, PartyChanges> = [
       'rink': 'archives:HolidayParty2010Rink(1).swf',
       'town': 'archives:RoomsTown-HolidayParty2010.swf'
     },
-    globalChanges: {
-      'content/map.swf': 'archives:HolidayParty2010Map-StadiumGames.swf'
-    },
+    map: 'archives:HolidayParty2010Map-StadiumGames.swf',
     localChanges: {
       'close_ups/christmasposter.swf': {
         'en': 'archives:HolidayParty2010ChristmasPoster.swf'
@@ -3120,9 +3120,7 @@ export const PARTIES: TemporaryUpdateTimeline<Party, PartyChanges> = [
           'rink': 'archives:HolidayParty2010Rink(2).swf',
           'town': 'archives:RoomsTown-HolidayParty2010_2.swf'
         },
-        globalChanges: {
-          'content/map.swf': 'archives:HolidayParty2010Map.swf'
-        }
+        map: 'archives:HolidayParty2010Map.swf'
       }
     ]
   }
