+++ src/server/game-data/updates.ts
@@ -144,3 +144,5 @@ export const AGENTCOM_RELEASE = '2008-11-24';
 export const GAME_UPGRADES = '2009-03-27';
 
 export const FIRE_CELEBRATION_START = '2009-11-20';
+
+export const FIRE_CONST_START = '2009-11-13';
